Move TaskList class to task package
Fixed javadoc for Command classes.
