+++ src/main/java/duke/Duke.java
@@ -1,6 +1,7 @@
 package duke;
 
 import duke.command.Command;
+import duke.task.TaskList;
 import duke.ui.Ui;
 
 /**
