+++ src/main/java/duke/task/TaskList.java
@@ -1,8 +1,9 @@
-package duke;
+package duke.task;
 
 import java.util.ArrayList;
 import java.util.stream.Collectors;
 
+import duke.DukeException;
 import duke.task.Deadline;
 import duke.task.Event;
 import duke.task.Task;
